Download several datasets given their URLs and destination path

This function enables to download datasets. In a data import workflow, this function is typically used after a call to the mf_get_url function. The output value of mf_get_url can be used as input of parameter df_to_dl of mf_download_data.

The download can the parallelized.

mf_download_data(
  df_to_dl,
  path = tempfile("tmp"),
  parallel = TRUE,
  credentials = NULL,
  source = "earthdata",
  verbose = TRUE,
  min_filesize = 5000
)

Arguments

df_to_dl

data.frame. Urls and destination files of dataset to download. Typically output of mf_get_url. See Details for the structure

path

string. Target folder for the data to download. Default : tempfile()

parallel

boolean. Parallelize the download ? Default to FALSE

credentials

vector string of length 2 with username and password. optional.

source

source. See details

verbose

boolean. optional. Verbose (default TRUE)

min_filesize

integer. minimum file size expected (in bites) for one file downloaded. If files downloaded are less that this value, the files will be downloaded again. Default 5000.

Value

a data.frame with the same structure of the input data.frame df_to_dl + columns providing details of the data downloaded. The additional columns are :

fileDl

Booloean (dataset downloaded or failure)

dlStatus

Download status : 1 = download ok ; 2 = download error ; 3 = dataset was already existing in destination file

fileSize

File size on disk

Details

Parameter df_to_dl must a data.frame with the following minimal structure :

url

URL of the source file (character string)

destfile

Destination file (character string)

Parameter source takes NULL as default value. Options are :

NULL : when no mf_login is required to download the data.

"earthdata" : to download data requiring a mf_login to EOSDIS Earthdata. To create an account go to : https://urs.earthdata.nasa.gov/