Precompute the parameter opt_param of the function odr_get_url

Precompute the parameter opt_param to further provide as input of the odr_get_url function. Useful to speed-up the overall processing time.

odr_get_opt_param(collection, roi, credentials = NULL, verbose = TRUE)

Arguments

collection
string. mandatory. Collection of interest (see details of odr_get_url).

roi
object of class sf or sfc. mandatory. Region of interest. Must be POLYGON-type geometry composed of one single feature.

credentials
vector string of length 2 with username and password. optional.

verbose
boolean. optional. Verbose (default TRUE)

Value

a list with the following named objects :

roiSpatialIndexBound

OPeNDAP indices for the spatial coordinates of the bounding box of the ROI (minLat, maxLat, minLon, maxLon)

availableVariables

Variables available for the collection of interest

roiSpatialBound

The spatial coordinates of the bounding box of the ROI expressed in the CRS of the collection

OpenDAPXVector

The X (longitude) vector

OpenDAPYVector

The Y (longitude) vector

OpenDAPtimeVector

The time vector, or NULL if the collection does not have a time vector

modis_tile

The MODIS tile(s) number(s) for the ROI or NULL if the collection is not MODIS

Details

When it is needed to loop the function odr_get_url over several time frames, it is advised to previously run the function odr_get_opt_param and provide the output as input opt_param parameter of the odr_get_url function. This will save much time, as internal parameters will be calculated only once.

Examples

require(sf)

#> Le chargement a nécessité le package : sf

#> Linking to GEOS 3.7.0, GDAL 2.3.2, PROJ 5.2.0

require(purrr)

#> Le chargement a nécessité le package : purrr

#> 
#> Attachement du package : ‘purrr’

#> The following object is masked from ‘package:testthat’:
#> 
#>     is_null

# Login to Earthdata
log <- odr_login(c(Sys.getenv("earthdata_un"),Sys.getenv("earthdata_pw")),source="earthdata")

#> Checking credentials...
#> Successfull login to earthdata 

# Get the optional parameters for the collection MOD11A1.006 and the roi :
roi <- st_as_sf(data.frame(
geom="POLYGON ((-5.82 9.54, -5.42 9.55, -5.41 8.84, -5.81 8.84, -5.82 9.54))"),
wkt="geom",crs = 4326)

opt_param_mod11a1 <- odr_get_opt_param("MOD11A1.006",roi)

#> Note : messages above ('although coordinates are longitude/latitude, st_intersection assumes that they are planar' and 'attribute variables are assumed to be spatially constant throughout all geometries') are not errors

#> although coordinates are longitude/latitude, st_intersection assumes that they are planar

#> Warning: attribute variables are assumed to be spatially constant throughout all geometries

#> although coordinates are longitude/latitude, st_intersection assumes that they are planar

#> Warning: attribute variables are assumed to be spatially constant throughout all geometries

str(opt_param_mod11a1)

#> List of 7
#>  $ roiSpatialIndexBound:List of 1
#>   ..$ 1: num [1:4] 55 140 512 560
#>  $ availableVariables  :'data.frame':	18 obs. of  6 variables:
#>   ..$ name                  : chr [1:18] "Latitude" "Longitude" "YDim" "XDim" ...
#>   ..$ long_name             : chr [1:18] "Latitude" "Longitude" "y coordinate of projection for grid MODIS_Grid_Daily_1km_LST" "x coordinate of projection for grid MODIS_Grid_Daily_1km_LST" ...
#>   ..$ units                 : chr [1:18] NA NA "meter" "meter" ...
#>   ..$ indices               : chr [1:18] "[YDim = 1200] [XDim = 1200]" "[YDim = 1200] [XDim = 1200]" "[YDim = 1200]" "[XDim = 1200]" ...
#>   ..$ all_info              : chr [1:18] "Array of 64 bit Reals [YDim = 0..1199][XDim = 0..1199]long_name: Latitude\nunits: degrees_north" "Array of 64 bit Reals [YDim = 0..1199][XDim = 0..1199]long_name: Longitude\nunits: degrees_east" "Array of 64 bit Reals [YDim = 0..1199]standard_name: projection_y_coordinate\nlong_name: y coordinate of projec"| __truncated__ "Array of 64 bit Reals [XDim = 0..1199]standard_name: projection_x_coordinate\nlong_name: x coordinate of projec"| __truncated__ ...
#>   ..$ extractable_w_opendapr: chr [1:18] "not extractable" "not extractable" "automatically extracted" "automatically extracted" ...
#>  $ roiSpatialBound     : NULL
#>  $ OpenDAPXVector      :List of 1
#>   ..$ 1: num [1:1200] -1111951 -1111024 -1110097 -1109171 -1108244 ...
#>  $ OpenDAPYVector      :List of 1
#>   ..$ 1: num [1:1200] 1111951 1111024 1110097 1109171 1108244 ...
#>  $ OpenDAPtimeVector   :List of 1
#>   ..$ 1: num [1:7243] 54 55 56 57 58 59 60 61 62 63 ...
#>  $ modis_tile          :List of 1
#>   ..$ 1: chr "h17v08"

# Now we can provide opt_param_mod11a1 as input parameter of the function odr_get_url().

time_ranges <- list(as.Date(c("2016-01-01","2016-01-31")),
                   as.Date(c("2017-01-01","2017-01-31")),
                   as.Date(c("2018-01-01","2018-01-31")),
                   as.Date(c("2019-01-01","2019-01-31")))

(urls_mod11a1 <- map(.x = time_ranges, ~odr_get_url(
 collection = "MOD11A1.006",
 variables = c("LST_Day_1km","LST_Night_1km","QC_Day","QC_Night"),
 roi = roi,
 time_range = .x,
 opt_param = opt_param_mod11a1)
))

#> Building the URLs...
#> OK
#> Building the URLs...
#> OK
#> Building the URLs...
#> OK
#> Building the URLs...
#> OK

#> [[1]]
#>   time_start                               name
#> 1 2016-01-01 MOD11A1.006.2016001_2016031.h17v08
#>                                                                                                                                                                                                                                                                                                                     url
#> 1 https://opendap.cr.usgs.gov/opendap/hyrax/MOD11A1.006/h17v08.ncml.nc4?MODIS_Grid_Daily_1km_LST_eos_cf_projection,LST_Day_1km[5736:5766][55:140][512:560],LST_Night_1km[5736:5766][55:140][512:560],QC_Day[5736:5766][55:140][512:560],QC_Night[5736:5766][55:140][512:560],time[5736:5766],YDim[55:140],XDim[512:560]
#>                                             destfile
#> 1 MOD11A1.006/MOD11A1.006.2016001_2016031.h17v08.nc4
#> 
#> [[2]]
#>   time_start                               name
#> 1 2017-01-01 MOD11A1.006.2017001_2017031.h17v08
#>                                                                                                                                                                                                                                                                                                                     url
#> 1 https://opendap.cr.usgs.gov/opendap/hyrax/MOD11A1.006/h17v08.ncml.nc4?MODIS_Grid_Daily_1km_LST_eos_cf_projection,LST_Day_1km[6093:6123][55:140][512:560],LST_Night_1km[6093:6123][55:140][512:560],QC_Day[6093:6123][55:140][512:560],QC_Night[6093:6123][55:140][512:560],time[6093:6123],YDim[55:140],XDim[512:560]
#>                                             destfile
#> 1 MOD11A1.006/MOD11A1.006.2017001_2017031.h17v08.nc4
#> 
#> [[3]]
#>   time_start                               name
#> 1 2018-01-01 MOD11A1.006.2018001_2018031.h17v08
#>                                                                                                                                                                                                                                                                                                                     url
#> 1 https://opendap.cr.usgs.gov/opendap/hyrax/MOD11A1.006/h17v08.ncml.nc4?MODIS_Grid_Daily_1km_LST_eos_cf_projection,LST_Day_1km[6458:6488][55:140][512:560],LST_Night_1km[6458:6488][55:140][512:560],QC_Day[6458:6488][55:140][512:560],QC_Night[6458:6488][55:140][512:560],time[6458:6488],YDim[55:140],XDim[512:560]
#>                                             destfile
#> 1 MOD11A1.006/MOD11A1.006.2018001_2018031.h17v08.nc4
#> 
#> [[4]]
#>   time_start                               name
#> 1 2019-01-01 MOD11A1.006.2019001_2019031.h17v08
#>                                                                                                                                                                                                                                                                                                                     url
#> 1 https://opendap.cr.usgs.gov/opendap/hyrax/MOD11A1.006/h17v08.ncml.nc4?MODIS_Grid_Daily_1km_LST_eos_cf_projection,LST_Day_1km[6823:6853][55:140][512:560],LST_Night_1km[6823:6853][55:140][512:560],QC_Day[6823:6853][55:140][512:560],QC_Night[6823:6853][55:140][512:560],time[6823:6853],YDim[55:140],XDim[512:560]
#>                                             destfile
#> 1 MOD11A1.006/MOD11A1.006.2019001_2019031.h17v08.nc4
#>