Import the time series in R as a SpatRaster object

Import a time series as a SpatRasterobject

mf_import_data(
  dir_path,
  collection_source,
  output_class = "SpatRaster",
  proj_epsg = "4326"
)

Arguments

dir_path

string character vector. mandatory. The path to the local directory where the data are stored.

collection_source

character string. mandatory. The collection source (one of "MODIS", "VIIRS", "GPM")

output_class

character string. Output object class. Currently only "SpatRaster" implemented.

proj_epsg

character string. EPSG of the desired projection for the output raster (default : 4326)

Examples

# \donttest{

require(sf)
require(magrittr)
require(terra)

username <- Sys.getenv("earthdata_un")
password <- Sys.getenv("earthdata_pw")
log <- mf_login(credentials = c(username,password))
#> Checking credentials...
#> Successfull login to earthdata 

roi_id <- "korhogo"
roi = st_as_sf(data.frame(
geom="POLYGON ((-5.82 9.54, -5.42 9.55, -5.41 8.84, -5.81 8.84, -5.82 9.54))"),
wkt="geom",crs = 4326)

time_range = as.Date(c("2017-01-01","2017-01-30"))

urls_mod11a1 <- mf_get_url(collection = "MOD11A1.061",
variables = c("LST_Day_1km","LST_Night_1km"),
roi = roi,
roi_id = roi_id,
time_range = time_range
)
#> Building the URLs...
#> Note : message 'attribute variables are assumed to be spatially constant throughout all geometries' is not an error
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
#> OK

res_dl <- mf_download_data(urls_mod11a1)
#> 1  datasets in total :  0  already downloaded and  1  datasets to download
#> Downloading the data...
#> Data were all properly downloaded

## import as terra::SpatRast

modis_ts <- mf_import_data(file.path(roi_id,"MOD11A1.061"), collection_source = "MODIS")

plot(mod11a1_rast)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'mod11a1_rast' not found

# }